chore: remove duplicate and unused API code from the agent crate
