+++ crates/agent/src/agent/agent_loop/model.rs
@@ -6,10 +6,6 @@ use std::sync::{
 use std::time::Duration;
 
 use futures::Stream;
-use serde::{
-    Deserialize,
-    Serialize,
-};
 use tokio::sync::mpsc;
 use tokio_stream::wrappers::ReceiverStream;
 use tokio_util::sync::CancellationToken;
@@ -27,7 +23,6 @@ use super::types::{
     Message,
     ToolSpec,
 };
-use crate::agent::rts::RtsModel;
 
 /// Represents a backend implementation for a converse stream compatible API.
 ///
@@ -48,59 +43,6 @@ pub trait Model: std::fmt::Debug + Send + Sync + 'static {
     /// associated with an implementation, useful for restoring a previous conversation.
     fn state(&self) -> Option<serde_json::Value> {
         None
-    }
-}
-
-/// The supported backends
-#[derive(Debug, Clone)]
-pub enum Models {
-    Rts(RtsModel),
-    Test(MockModel),
-}
-
-impl Models {
-    pub fn state(&self) -> ModelsState {
-        match self {
-            Models::Rts(v) => ModelsState::Rts {
-                conversation_id: Some(v.conversation_id().to_string()),
-                model_id: v.model_id().map(String::from),
-            },
-            Models::Test(_) => ModelsState::Test,
-        }
-    }
-}
-
-/// A serializable representation of the state contained within [Models].
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub enum ModelsState {
-    Rts {
-        conversation_id: Option<String>,
-        model_id: Option<String>,
-    },
-    Test,
-}
-
-impl Default for ModelsState {
-    fn default() -> Self {
-        Self::Rts {
-            conversation_id: None,
-            model_id: None,
-        }
-    }
-}
-
-impl Model for Models {
-    fn stream(
-        &self,
-        messages: Vec<Message>,
-        tool_specs: Option<Vec<ToolSpec>>,
-        system_prompt: Option<String>,
-        cancel_token: CancellationToken,
-    ) -> Pin<Box<dyn Stream<Item = StreamResult> + Send + 'static>> {
-        match self {
-            Models::Rts(rts_model) => rts_model.stream(messages, tool_specs, system_prompt, cancel_token),
-            Models::Test(test_model) => test_model.stream(messages, tool_specs, system_prompt, cancel_token),
-        }
     }
 }
