--- crates/agent/src/agent/rts/util.rs
@@ -1,56 +1,0 @@
-use aws_smithy_types::{
-    Document,
-    Number as SmithyNumber,
-};
-
-pub fn serde_value_to_document(value: serde_json::Value) -> Document {
-    match value {
-        serde_json::Value::Null => Document::Null,
-        serde_json::Value::Bool(bool) => Document::Bool(bool),
-        serde_json::Value::Number(number) => {
-            if let Some(num) = number.as_u64() {
-                Document::Number(SmithyNumber::PosInt(num))
-            } else if number.as_i64().is_some_and(|n| n < 0) {
-                Document::Number(SmithyNumber::NegInt(number.as_i64().unwrap()))
-            } else {
-                Document::Number(SmithyNumber::Float(number.as_f64().unwrap_or_default()))
-            }
-        },
-        serde_json::Value::String(string) => Document::String(string),
-        serde_json::Value::Array(vec) => {
-            Document::Array(vec.clone().into_iter().map(serde_value_to_document).collect::<_>())
-        },
-        serde_json::Value::Object(map) => Document::Object(
-            map.into_iter()
-                .map(|(k, v)| (k, serde_value_to_document(v)))
-                .collect::<_>(),
-        ),
-    }
-}
-
-pub fn document_to_serde_value(value: Document) -> serde_json::Value {
-    use serde_json::Value;
-    match value {
-        Document::Object(map) => Value::Object(
-            map.into_iter()
-                .map(|(k, v)| (k, document_to_serde_value(v)))
-                .collect::<_>(),
-        ),
-        Document::Array(vec) => Value::Array(vec.clone().into_iter().map(document_to_serde_value).collect::<_>()),
-        Document::Number(number) => {
-            if let Ok(v) = TryInto::<u64>::try_into(number) {
-                Value::Number(v.into())
-            } else if let Ok(v) = TryInto::<i64>::try_into(number) {
-                Value::Number(v.into())
-            } else {
-                Value::Number(
-                    serde_json::Number::from_f64(number.to_f64_lossy())
-                        .unwrap_or(serde_json::Number::from_f64(0.0).expect("converting from 0.0 will not fail")),
-                )
-            }
-        },
-        Document::String(s) => serde_json::Value::String(s),
-        Document::Bool(b) => serde_json::Value::Bool(b),
-        Document::Null => serde_json::Value::Null,
-    }
-}
