--- crates/agent/src/api_client/send_message_output.rs
@@ -1,45 +1,0 @@
-use aws_types::request_id::RequestId;
-
-use crate::api_client::ApiClientError;
-use crate::api_client::model::ChatResponseStream;
-
-#[derive(Debug)]
-pub enum SendMessageOutput {
-    Codewhisperer(
-        amzn_codewhisperer_streaming_client::operation::generate_assistant_response::GenerateAssistantResponseOutput,
-    ),
-    QDeveloper(amzn_qdeveloper_streaming_client::operation::send_message::SendMessageOutput),
-    Mock(Vec<ChatResponseStream>),
-}
-
-impl SendMessageOutput {
-    pub fn request_id(&self) -> Option<&str> {
-        match self {
-            SendMessageOutput::Codewhisperer(output) => output.request_id(),
-            SendMessageOutput::QDeveloper(output) => output.request_id(),
-            SendMessageOutput::Mock(_) => None,
-        }
-    }
-
-    pub async fn recv(&mut self) -> Result<Option<ChatResponseStream>, ApiClientError> {
-        match self {
-            SendMessageOutput::Codewhisperer(output) => Ok(output
-                .generate_assistant_response_response
-                .recv()
-                .await?
-                .map(|s| s.into())),
-            SendMessageOutput::QDeveloper(output) => Ok(output.send_message_response.recv().await?.map(|s| s.into())),
-            SendMessageOutput::Mock(vec) => Ok(vec.pop()),
-        }
-    }
-}
-
-impl RequestId for SendMessageOutput {
-    fn request_id(&self) -> Option<&str> {
-        match self {
-            SendMessageOutput::Codewhisperer(output) => output.request_id(),
-            SendMessageOutput::QDeveloper(output) => output.request_id(),
-            SendMessageOutput::Mock(_) => Some("<mock-request-id>"),
-        }
-    }
-}
