--- crates/agent/src/auth/mod.rs
@@ -1,56 +1,0 @@
-pub mod builder_id;
-mod consts;
-
-use aws_sdk_ssooidc::error::SdkError;
-use aws_sdk_ssooidc::operation::create_token::CreateTokenError;
-use aws_sdk_ssooidc::operation::register_client::RegisterClientError;
-use aws_sdk_ssooidc::operation::start_device_authorization::StartDeviceAuthorizationError;
-use thiserror::Error;
-
-use crate::agent::util::error::UtilError;
-
-#[derive(Debug, Error)]
-pub enum AuthError {
-    #[error(transparent)]
-    Ssooidc(Box<aws_sdk_ssooidc::Error>),
-    #[error(transparent)]
-    SdkRegisterClient(Box<SdkError<RegisterClientError>>),
-    #[error(transparent)]
-    SdkCreateToken(Box<SdkError<CreateTokenError>>),
-    #[error(transparent)]
-    SdkStartDeviceAuthorization(Box<SdkError<StartDeviceAuthorizationError>>),
-    #[error(transparent)]
-    Io(#[from] std::io::Error),
-    #[error(transparent)]
-    TimeComponentRange(#[from] time::error::ComponentRange),
-    #[error(transparent)]
-    SerdeJson(#[from] serde_json::Error),
-    #[error(transparent)]
-    Util(#[from] UtilError),
-    #[error("No token")]
-    NoToken,
-}
-
-impl From<aws_sdk_ssooidc::Error> for AuthError {
-    fn from(value: aws_sdk_ssooidc::Error) -> Self {
-        Self::Ssooidc(Box::new(value))
-    }
-}
-
-impl From<SdkError<RegisterClientError>> for AuthError {
-    fn from(value: SdkError<RegisterClientError>) -> Self {
-        Self::SdkRegisterClient(Box::new(value))
-    }
-}
-
-impl From<SdkError<CreateTokenError>> for AuthError {
-    fn from(value: SdkError<CreateTokenError>) -> Self {
-        Self::SdkCreateToken(Box::new(value))
-    }
-}
-
-impl From<SdkError<StartDeviceAuthorizationError>> for AuthError {
-    fn from(value: SdkError<StartDeviceAuthorizationError>) -> Self {
-        Self::SdkStartDeviceAuthorization(Box::new(value))
-    }
-}
