--- crates/agent/src/database/sqlite_migrations/000_create_migration_auth_state_tables.sql
@@ -1,14 +1,0 @@
-CREATE TABLE IF NOT EXISTS migrations (
-    version INTEGER PRIMARY KEY,
-    migration_time INTEGER NOT NULL
-);
-
-CREATE TABLE IF NOT EXISTS auth_kv (
-    key TEXT PRIMARY KEY,
-    value TEXT
-);
-
-CREATE TABLE IF NOT EXISTS state (
-    key TEXT PRIMARY KEY,
-    value TEXT
-);
