+++ crates/agent/src/lib.rs
@@ -1,8 +1,4 @@
 mod agent;
-pub mod api_client;
-mod auth;
-mod aws_common;
-mod database;
 
 // TODO - probably should fix imports after removing all of the duplicated api client and database
 // code.
